+++ R/generics.R
@@ -66,7 +66,8 @@ setGeneric("identifiers", function(compounds) standardGeneric("identifiers"))
 setGeneric("addFormulaScoring", function(compounds, formulas, updateScore = FALSE,
                                          formulaScoreWeight = 1) standardGeneric("addFormulaScoring"))
 setGeneric("plotScores", function(obj, index, groupName, normalizeScores = "max",
-                                  excludeNormScores = NULL, onlyUsed = TRUE, useGGPlot2 = FALSE) standardGeneric("plotScores"))
+                                  excludeNormScores = c("score", "individualMoNAScore"), onlyUsed = TRUE,
+                                  useGGPlot2 = FALSE) standardGeneric("plotScores"))
 setGeneric("settings", function(compoundsMF) standardGeneric("settings"))
 
 ### clustering
@@ -279,17 +280,20 @@ setGeneric("checkChromatograms", function(fGroups, mzWindow = 0.005, enabledFGro
 setGeneric("compoundViewer", function(fGroups, MSPeakLists, compounds) standardGeneric("compoundViewer"))
 setGeneric("reportCSV", function(fGroups, path = "report", reportFGroupsAsRows = TRUE, reportFGroupsAnalysisInfo = TRUE,
                                  reportFGroupsRetMz = TRUE, reportFeatures = FALSE, formulas = NULL,
-                                 compounds = NULL, compoundNormalizeScores = "max", compoundExclNormScores = "score",
+                                 compounds = NULL, compoundNormalizeScores = "max",
+                                 compoundExclNormScores = c("score", "individualMoNAScore"),
                                  compsCluster = NULL, components = NULL,
                                  retMin = TRUE, clearPath = FALSE) standardGeneric("reportCSV"))
 setGeneric("reportPDF", function(fGroups, path = "report", reportFGroups = TRUE,
                                  formulas = NULL, reportFormulaSpectra = TRUE, compounds = NULL, compoundNormalizeScores = "max",
-                                 compoundExclNormScores = "score", compoundOnlyUsedScorings = TRUE, compoundTopMost = 5,
+                                 compoundExclNormScores = c("score", "individualMoNAScore"),
+                                 compoundOnlyUsedScorings = TRUE, compoundTopMost = 5,
                                  compsCluster = NULL, components = NULL, MSPeakLists = NULL, retMin = TRUE,
                                  EICGrid = c(2, 1), EICRtWindow = 20, EICMzWindow = 0.005, EICTopMost = NULL,
                                  EICOnlyPresent = TRUE, clearPath = FALSE) standardGeneric("reportPDF"))
 setGeneric("reportMD", function(fGroups, path = "report", reportPlots = c("chord", "venn", "upset", "eics", "formulas"),
-                                formulas = NULL, compounds = NULL, compoundNormalizeScores = "max", compoundExclNormScores = "score",
+                                formulas = NULL, compounds = NULL, compoundNormalizeScores = "max",
+                                compoundExclNormScores = c("score", "individualMoNAScore"),
                                 compoundOnlyUsedScorings = TRUE, compoundTopMost = 5, compsCluster = NULL,
                                 includeMFWebLinks = "compounds", components = NULL, interactiveHeat = FALSE,
                                 MSPeakLists = NULL, retMin = TRUE, EICRtWindow = 20, EICMzWindow = 0.005,